Update Jekyll config:
Changed name of the site and fixed the url to point to the site, not
my GitHub page. I changed the base url to remove the reference to
jekyll-uno.

To allow posts written with the date set in the future, this includes
posts marked as todays date, I had a problems until I figured this out.
I set future: true